‼️ BREAKING: Change from template to a CLI tool
