--- config/keys.js
@@ -1,3 +1,0 @@
-module.exports = {
-    MongoURI: '[YOUR CONNECTION STRING HERE]'
-}
