+++ templates/ejs-template/views/docs/index.ejs
