+++ templates/node-mongo-template/.env
