+++ templates/node-mongo-template/api/api.js
