+++ templates/node-mongo-template/controllers/toycar-controller.js
