+++ templates/node-mongo-template/models/ToyCar.js
