+++ templates/node-mongo-template/routes/index.js
